Migrate Faust Interpreter effect wwise plugin to source wwise plugin. Deleted the previous wwise project and created a new one. Now the interpreter is implemented as a Source plugin.
