--- Faust_Compiler_Test_Plugin/.vscode/launch.json
@@ -1,19 +1,0 @@
-{
-	// Use IntelliSense to learn about possible attributes.
-	// Hover to view descriptions of existing attributes.
-	// For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-	"version": "0.2.0",
-	"configurations": [
-		{
-			"name": "Wwise Authoring",
-			"type": "cppvsdbg",
-			"request": "launch",
-			"program": "${env:WWISEROOT}/Authoring/x64/Release/bin/Wwise.exe",
-			"args": [],
-			"stopAtEntry": false,
-			"cwd": "${workspaceFolder}",
-			"environment": [],
-			"console": "integratedTerminal"
-		}
-	]
-}
