--- Faust_Compiler_Test_Plugin/.vscode/settings.json
@@ -1,12 +1,0 @@
-{
-	"files.exclude": {
-		"**/*.sln": true,
-		"**/*.vcxproj": true,
-		"**/*.vcxproj.filters": true,
-		"**/*.vcxproj.user": true,
-		"**/*.aps": true,
-		"**/*.mk": true,
-		"**/*.tar.xz": true,
-		"Output": true
-	}
-}
