--- Faust_Compiler_Test_Plugin/SoundEnginePlugin/Faust_Compiler_Test_PluginFXParams.h
@@ -1,77 +1,0 @@
-/*******************************************************************************
-The content of this file includes portions of the AUDIOKINETIC Wwise Technology
-released in source code form as part of the SDK installer package.
-
-Commercial License Usage
-
-Licensees holding valid commercial licenses to the AUDIOKINETIC Wwise Technology
-may use this file in accordance with the end user license agreement provided
-with the software or, alternatively, in accordance with the terms contained in a
-written agreement between you and Audiokinetic Inc.
-
-Apache License Usage
-
-Alternatively, this file may be used under the Apache License, Version 2.0 (the
-"Apache License"); you may not use this file except in compliance with the
-Apache License. You may obtain a copy of the Apache License at
-http://www.apache.org/licenses/LICENSE-2.0.
-
-Unless required by applicable law or agreed to in writing, software distributed
-under the Apache License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
-OR CONDITIONS OF ANY KIND, either express or implied. See the Apache License for
-the specific language governing permissions and limitations under the License.
-
-  Copyright (c) 2006-2022 Audiokinetic Inc.
-*******************************************************************************/
-
-#ifndef Faust_Compiler_Test_PluginFXParams_H
-#define Faust_Compiler_Test_PluginFXParams_H
-
-#include <AK/SoundEngine/Common/IAkPlugin.h>
-#include <AK/Plugin/PluginServices/AkFXParameterChangeHandler.h>
-
-// Add parameters IDs here, those IDs should map to the AudioEnginePropertyID
-// attributes in the xml property definition.
-static const AkPluginParamID PARAM_PLACEHOLDER_ID = 0;
-static const AkUInt32 NUM_PARAMS = 1;
-
-struct Faust_Compiler_Test_PluginRTPCParams
-{
-    AkReal32 fPlaceholder;
-};
-
-struct Faust_Compiler_Test_PluginNonRTPCParams
-{
-};
-
-struct Faust_Compiler_Test_PluginFXParams
-    : public AK::IAkPluginParam
-{
-    Faust_Compiler_Test_PluginFXParams();
-    Faust_Compiler_Test_PluginFXParams(const Faust_Compiler_Test_PluginFXParams& in_rParams);
-
-    ~Faust_Compiler_Test_PluginFXParams();
-
-    /// Create a duplicate of the parameter node instance in its current state.
-    IAkPluginParam* Clone(AK::IAkPluginMemAlloc* in_pAllocator) override;
-
-    /// Initialize the plug-in parameter node interface.
-    /// Initializes the internal parameter structure to default values or with the provided parameter block if it is valid.
-    AKRESULT Init(AK::IAkPluginMemAlloc* in_pAllocator, const void* in_pParamsBlock, AkUInt32 in_ulBlockSize) override;
-
-    /// Called by the sound engine when a parameter node is terminated.
-    AKRESULT Term(AK::IAkPluginMemAlloc* in_pAllocator) override;
-
-    /// Set all plug-in parameters at once using a parameter block.
-    AKRESULT SetParamsBlock(const void* in_pParamsBlock, AkUInt32 in_ulBlockSize) override;
-
-    /// Update a single parameter at a time and perform the necessary actions on the parameter changes.
-    AKRESULT SetParam(AkPluginParamID in_paramID, const void* in_pValue, AkUInt32 in_ulParamSize) override;
-
-    AK::AkFXParameterChangeHandler<NUM_PARAMS> m_paramChangeHandler;
-
-    Faust_Compiler_Test_PluginRTPCParams RTPC;
-    Faust_Compiler_Test_PluginNonRTPCParams NonRTPC;
-};
-
-#endif // Faust_Compiler_Test_PluginFXParams_H
