--- Faust_Compiler_Test_Plugin/WwisePlugin/Faust_Compiler_Test_Plugin.def
@@ -1,3 +1,0 @@
-LIBRARY "Faust_Compiler_Test_Plugin"
-
-EXPORTS
