--- Faust_Compiler_Test_Plugin/WwisePlugin/res/Md/en/Placeholder.md
@@ -1,7 +1,0 @@
-##Placeholder Parameter
-
-The description of what this parameter does...
-
-**Note**: Additional note goes here...
-
-Units: x <br/>
