+++ Faust_Interpreter_Test_Plugin/WwisePlugin/Faust.rc
