+++ Faust_Interpreter_Test_Plugin/WwisePlugin/Faust_Interpreter_Test_Plugin.def
@@ -1,0 +1,3 @@
+LIBRARY "Faust_Interpreter_Test_Plugin"
+
+EXPORTS
