+++ Faust_Interpreter_Test_Plugin/WwisePlugin/res/Md/en/Duration.md
@@ -1,0 +1,7 @@
+## Placeholder Parameter
+
+The description of what this parameter does.
+
+**Note**: Additional note goes here.
+
+Units: x <br/>
