+++ Faust_Interpreter_Test_Plugin/WwisePlugin/res/Md/ja/Duration.md
@@ -1,0 +1,7 @@
+## プレースホルダパラメータ
+
+このパラメータの機能の説明。
+
+**注**：ここに追加のメモが記載されます。
+
+単位：x <br/>
