+++ Faust_Interpreter_Test_Plugin/WwisePlugin/res/Md/ko/Duration.md
@@ -1,0 +1,7 @@
+## 자리 표시자 매개 변수
+
+이 매개 변수에 대한 설명
+
+**참고** : 추가 메모가 여기에 옵니다.
+
+단위 : x <br/>
