+++ Faust_Interpreter_Test_Plugin/WwisePlugin/res/Md/zh/Duration.md
@@ -1,0 +1,7 @@
+## 占位符参数
+
+该参数的功能说明。
+
+**注意**： 这部分为附加说明 。
+
+单位：x <br/>
