+++ Faust_Interpreter_Test_Plugin/WwisePlugin/resource.h
