+++ README.md
@@ -4,4 +4,4 @@ Faust/Wwise tools:
 
 - Faust_Test_Plugin is a preliminary work to develop a `faust2wwise` tool to *statically* compile Faust DSP code in WWise modules
 
-- Faust_Compiler_Test_Plugin is a preliminary work to develop an integrated Faust interpreter as a Wwise plugin using [libfaust + Interp](https://faustdoc.grame.fr/manual/embedding/)
+- Faust_Interpreter_Test_Plugin is a preliminary work to develop an integrated Faust interpreter as a Wwise (**Source**) Plugin using [libfaust + Interpreter backend](https://faustdoc.grame.fr/manual/embedding/)
